requirements.txt
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
Deprecated==1.2.12
Flask==2.0.1
Flask-Dance==5.0.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
oauthlib==3.1.1
opencv-python==4.5.2.54
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
python-dotenv==0.18.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.6
URLObject==2.4.3
Werkzeug==2.0.1
wrapt==1.12.1
gunicorn
Pillow~=8.2.0
Flask-SocketIO==2.9.1
